Rewritten Huffman codec with L1d-cache-aware decode LUT sizing
