+++ Cargo.toml
@@ -37,6 +37,9 @@ zune-inflate = { version = "^0.2.54", default-features = false, features = ["zli
 pulp = "^0.22.3"                # used for runtime SIMD dispatch
 num-complex = "0.4.6"          # pins pulps own dependency, pulp wants 0.4.4 but calls for Complex::ZERO/ONE/I; consts added in 0.4.6
 
+[target.'cfg(any(target_arch = "x86", target_arch = "x86_64"))'.dependencies]
+raw-cpuid = { version = "11.6.0", default-features = false }  # detect L1d cache size to pick the Huffman decode LUT width
+
 [dev-dependencies]
 bencher = "0.1.5"
 walkdir = "2.3.2"         # automatically test things for all files in a directory
@@ -61,6 +64,14 @@ sse2-tests = []
 # Meaningless under emulation (timings won't reflect real hardware)
 # `cargo bench --features simd-benches --bench dct`
 simd-benches = []
+
+# Override the AC-Huffman decode prefix-LUT size (huffman.rs' LUT_BITS).
+# Bigger trades fewer long-code fallbacks for more L1d pressure per DWA
+# chunk; without one of these pinned, x86/x86_64 picks it at runtime from
+# the detected L1d size instead. Mutually exclusive.
+huffman-lut-12 = []
+huffman-lut-13 = []
+huffman-lut-14 = []
 
 [[example]]
 name = "7_write_raw_blocks"
